Segregated some things into modules, added FK pragma for SQLITE
